docs: factorize doc
