+++ doc/src/app/geops-mobility-search/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import GeopsMobilitySearchDoc from "../components/GeopsMobilitySearchDoc";
+
+export default function GeopsMobilitySearch() {
+  return <GeopsMobilitySearchDoc />;
+}
