+++ doc/src/app/geops-mobility/page.tsx
@@ -1,0 +1,6 @@
+"use client";
+import GeopsMobilityDoc from "../components/GeopsMobilityDoc";
+
+export default function GeopsMobility() {
+  return <GeopsMobilityDoc />;
+}
