+++ doc/src/geops-ui.ts
@@ -1,0 +1,3 @@
+"use client";
+// @ts-expect-error - required for module declaration
+export { Footer, geopsTheme, Header } from "@geops/geops-ui";
